Make this a Maven project.
