+++ .gitignore
@@ -1,1 +1,3 @@
-bin/
+*.class
+target/classes/
+target/maven-archiver/
