+++ pom.xml
@@ -1,0 +1,56 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+	<modelVersion>4.0.0</modelVersion>
+	<groupId>nl.mightydev.lumberjack</groupId>
+	<artifactId>lumberjack</artifactId>
+	<version>CCY-1.0</version>
+  <name>Lumberjack</name>
+  <description>Minecraft Lumberjack Bukkit Plugin</description>
+  <build>
+    <finalName>Lumberjack</finalName>
+    <sourceDirectory>${basedir}/src/main/java</sourceDirectory>
+    <resources>
+		<resource>
+			<targetPath>.</targetPath>
+			<filtering>true</filtering>
+			<directory>${basedir}/src/main/resources/</directory>
+			<includes>
+				<include>*.yml</include>
+			</includes>
+		</resource>
+    </resources>
+	<plugins>
+		<plugin>
+			<artifactId>maven-compiler-plugin</artifactId>
+			<version>2.3.2</version>
+			<configuration>
+				<source>1.6</source>
+				<target>1.6</target>
+			</configuration>
+		</plugin>
+    </plugins>
+  </build>
+	<repositories>
+		<repository>
+			<snapshots>
+				<enabled>true</enabled>
+				<updatePolicy>always</updatePolicy>
+				<checksumPolicy>fail</checksumPolicy>
+			</snapshots>
+			<id>bukkit-repo</id>
+			<url>http://repo.bukkit.org/content/groups/public/</url>
+		</repository>
+	</repositories>
+    <dependencies>
+        <dependency>
+            <groupId>org.bukkit</groupId>
+            <artifactId>bukkit</artifactId>
+            <version>LATEST</version>
+            <type>jar</type>
+            <scope>compile</scope>
+        </dependency>
+    </dependencies>
+    <properties>
+        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+    </properties>
+</project>
